Move vehicles to their own namespace. Remove unused imports
