+++ AutoAuctionProjekt/Classes/Vehicles/Bus.cs
@@ -1,8 +1,6 @@
 ﻿using System;
-using System.Collections.Generic;
-using System.Text;
 
-namespace AutoAuctionProjekt.Classes
+namespace AutoAuctionProjekt.Classes.Vehicles
 {
     class Bus : HeavyVehicle
     {
