+++ AutoAuctionProjekt/Program.cs
@@ -1,8 +1,7 @@
 ﻿using System;
 using System.Collections.Generic;
-using System.Data.SqlClient;
-using System.Text;
 using AutoAuctionProjekt.Classes;
+using AutoAuctionProjekt.Classes.Vehicles;
 
 namespace AutoAuctionProjekt
 {
@@ -13,12 +12,12 @@ namespace AutoAuctionProjekt
             //AuctionHouse objects init
             #region init car objects
             PersonalCar.TrunkDimentionsStruct td = new PersonalCar.TrunkDimentionsStruct(14.0, 10.0, 16.0);
-            HeavyVehicle.VehicleDimensionsStruct vd = new HeavyVehicle.VehicleDimensionsStruct(214.0, 2.59, 12.9);
+            VehicleDimensionsStruct vd = new(214.0, 2.59, 12.9);
 
-            PrivatePersonalCar privateCar1 = new ( "Some car brand", 300.0, "DF12745", 2009, 10000M, false, 10.0, 20.0, Vehicle.FuelTypeEnum.Diesel, 3, td, true);
-            PrivatePersonalCar privateCar2 = new ("Another car brand", 300.0, "DF12345", 2020, 12000M, true, 10.0, 20.0, Vehicle.FuelTypeEnum.Benzin, 5, td, false);
-            ProfessionalPersonalCar professionalCar = new ("Suzuki Swift", 500.0, "XY12345", 2012, 10000M, 10.0, 20.0, Vehicle.FuelTypeEnum.Benzin, 2, td, true, 400.0);
-            Bus bus = new Bus("City bus", 800.0, "HE24745", 2012, 30000M, true, 10.0, 15.0, Vehicle.FuelTypeEnum.Diesel, vd, 24, 10, true);
+            PrivatePersonalCar privateCar1 = new ( "Some car brand", 300.0, "DF12745", 2009, 10000M, false, 10.0, 20.0, FuelTypeEnum.Diesel, 3, td, true);
+            PrivatePersonalCar privateCar2 = new ("Another car brand", 300.0, "DF12345", 2020, 12000M, true, 10.0, 20.0, FuelTypeEnum.Benzin, 5, td, false);
+            ProfessionalPersonalCar professionalCar = new ("Suzuki Swift", 500.0, "XY12345", 2012, 10000M, 10.0, 20.0, FuelTypeEnum.Benzin, 2, td, true, 400.0);
+            Bus bus = new Bus("City bus", 800.0, "HE24745", 2012, 30000M, true, 10.0, 15.0, FuelTypeEnum.Diesel, vd, 24, 10, true);
             #endregion
 
             #region init user objects
